[Ingest Manager] Fix agent status filtering (#66787)
