+++ x-pack/plugins/ingest_manager/server/services/agents/status.ts
@@ -7,14 +7,8 @@
 import { SavedObjectsClientContract } from 'src/core/server';
 import { getAgent, listAgents } from './crud';
 import { AGENT_EVENT_SAVED_OBJECT_TYPE, AGENT_SAVED_OBJECT_TYPE } from '../../constants';
-import { AgentStatus, Agent } from '../../types';
+import { AgentStatus } from '../../types';
 
-import {
-  AGENT_POLLING_THRESHOLD_MS,
-  AGENT_TYPE_PERMANENT,
-  AGENT_TYPE_TEMPORARY,
-  AGENT_TYPE_EPHEMERAL,
-} from '../../constants';
 import { AgentStatusKueryHelper } from '../../../common/services';
 
 export async function getAgentStatusById(
@@ -22,39 +16,10 @@ export async function getAgentStatusById(
   agentId: string
 ): Promise<AgentStatus> {
   const agent = await getAgent(soClient, agentId);
-  return getAgentStatus(agent);
+  return AgentStatusKueryHelper.getAgentStatus(agent);
 }
 
-export function getAgentStatus(agent: Agent, now: number = Date.now()): AgentStatus {
-  const { type, last_checkin: lastCheckIn } = agent;
-  const msLastCheckIn = new Date(lastCheckIn || 0).getTime();
-  const msSinceLastCheckIn = new Date().getTime() - msLastCheckIn;
-  const intervalsSinceLastCheckIn = Math.floor(msSinceLastCheckIn / AGENT_POLLING_THRESHOLD_MS);
-  if (!agent.active) {
-    return 'inactive';
-  }
-  if (agent.current_error_events.length > 0) {
-    return 'error';
-  }
-  switch (type) {
-    case AGENT_TYPE_PERMANENT:
-      if (intervalsSinceLastCheckIn >= 4) {
-        return 'error';
-      }
-      if (intervalsSinceLastCheckIn >= 2) {
-        return 'warning';
-      }
-    case AGENT_TYPE_TEMPORARY:
-      if (intervalsSinceLastCheckIn >= 3) {
-        return 'offline';
-      }
-    case AGENT_TYPE_EPHEMERAL:
-      if (intervalsSinceLastCheckIn >= 3) {
-        return 'inactive';
-      }
-  }
-  return 'online';
-}
+export const getAgentStatus = AgentStatusKueryHelper.getAgentStatus;
 
 export async function getAgentStatusForConfig(
   soClient: SavedObjectsClientContract,
